Move litter to treatment
