+++ addons/medical_treatment/CfgVehicles.hpp
@@ -1,0 +1,44 @@
+class CfgVehicles {
+    // Medical litter classes
+    class Thing;
+    class ACE_MedicalLitterBase: Thing {
+        scope = 1;
+        scopeCurator = 0;
+        displayName = " ";
+        destrType = "DestructNo";
+        model = QPATHTOF(data\littergeneric.p3d);
+    };
+    class ACE_MedicalLitter_clean: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_clean.p3d);
+    };
+    class ACE_MedicalLitter_bandage1: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_bandages1.p3d);
+    };
+    class ACE_MedicalLitter_bandage2: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_bandages2.p3d);
+    };
+    class ACE_MedicalLitter_bandage3: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_bandages3.p3d);
+    };
+    class ACE_MedicalLitter_packingBandage: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_packingBandage.p3d);
+    };
+    class ACE_MedicalLitter_gloves: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_gloves.p3d);
+    };
+    class ACE_MedicalLitter_adenosine: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_adenosine.p3d);
+    };
+    class ACE_MedicalLitter_atropine: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_atropine.p3d);
+    };
+    class ACE_MedicalLitter_epinephrine: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_epinephrine.p3d);
+    };
+    class ACE_MedicalLitter_morphine: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_morphine.p3d);
+    };
+    class ACE_MedicalLitter_QuickClot: ACE_MedicalLitterBase {
+        model = QPATHTOF(data\littergeneric_Quikclot.p3d);
+    };
+};
