+++ addons/medical_treatment/config.cpp
@@ -14,7 +14,8 @@ class CfgPatches {
     };
 };
 
-#include "CfgEventHandlers.hpp"
 #include "ACE_Medical_Treatment.hpp"
 #include "ACE_Medical_Treatment_Actions.hpp"
 #include "ACE_Settings.hpp"
+#include "CfgEventHandlers.hpp"
+#include "CfgVehicles.hpp"
