+++ addons/medical_treatment/data/littergeneric.p3d
