+++ addons/medical_treatment/data/littergeneric_Quikclot.p3d
