+++ addons/medical_treatment/data/littergeneric_adenosine.p3d
