+++ addons/medical_treatment/data/littergeneric_atropine.p3d
