+++ addons/medical_treatment/data/littergeneric_bandages1.p3d
