+++ addons/medical_treatment/data/littergeneric_bandages2.p3d
