+++ addons/medical_treatment/data/littergeneric_bandages3.p3d
