+++ addons/medical_treatment/data/littergeneric_clean.p3d
