+++ addons/medical_treatment/data/littergeneric_epinephrine.p3d
