+++ addons/medical_treatment/data/littergeneric_gloves.p3d
