+++ addons/medical_treatment/data/littergeneric_gloves.rvmat
