+++ addons/medical_treatment/data/littergeneric_gloves_co.paa
