+++ addons/medical_treatment/data/littergeneric_morphine.p3d
