+++ addons/medical_treatment/data/littergeneric_packingbandage.p3d
