+++ addons/medical_treatment/data/littergeneric_packingbandage.rvmat
